// src/enums/RoomEnum.ts
export enum enumRoomType {
    SingleBed = "Single Bed",
    DoubleBed = "Double Bed",
    DoubleSuperior = "Double Superior",
    Suite = "Suite"
}

export enum enumAmenities {
    TV = "TV",
    Fridge = "Fridge",
    Bathtub = "Bathtub",
    SeaView = "Sea View",
    AirConditioner = "Air Conditioner"
}

export enum enumRoomStatus {
    Available = "Available",
    Booked = "Booked"
}